[commons] refactor ColorUtil to commons
