+++ bundles/org.smarthomej.commons/src/main/java/org/smarthomej/commons/util/ColorUtil.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.smarthomej.binding.deconz.internal;
+package org.smarthomej.commons.util;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.openhab.core.library.types.HSBType;
